Refactor & Linter errors
Ignore no-use-before-define-rule for function definitions, since they
are hoisted anyway and readability suffers a lot when reversing function
order from least-significant to most-significant.
