+++ src/util/flat.js
@@ -1,0 +1,8 @@
+// Thanks
+// https://stackoverflow.com/a/15030117/1870317
+
+export default function flatten(arr) {
+  return arr.reduce((flat, toFlatten) => (
+    flat.concat(Array.isArray(toFlatten) ? flatten(toFlatten) : toFlatten)
+  ), []);
+}
